Cleaned up UserAttributes model
